Modifica tipi double a decimal per prezzi e tasse
Il namespace `Api.Sales.Controllers` nel file `ProductController.cs` è stato modificato per utilizzare il tipo `decimal` al posto di `double` per i prezzi degli articoli (`ItemPrice`).

Nel namespace `SALESbe.Controllers` nel file `PurchaseController.cs`, il tipo `double` è stato sostituito con `decimal` per le variabili relative alle tasse e ai prezzi degli articoli. Inoltre, la funzione `RoundNumbersUp` è stata aggiornata per lavorare con `decimal` anziché `double`.

Nel file `RequestPurchaseDto.cs` del namespace `Api.Sales.Models.DTOs.Requests`, la proprietà `TotalAmount` è stata cambiata da `double` a `decimal`.

Nel file `ResponseProductDto.cs`, la proprietà `ItemPrice` è stata cambiata da `double` a `decimal` e il costruttore è stato aggiornato di conseguenza.

Nel file `ResponsePurchaseDto.cs` del namespace `Api.Sales.Models.DTOs.Responses`, le proprietà `SalesTaxes` e `Total` sono state cambiate da `double` a `decimal`.

Nel file `Product.cs`, le proprietà `ItemPrice` e `TotalPrice` sono state cambiate da `double` a `decimal` e il costruttore è stato aggiornato di conseguenza.

+++ SALESbe/Models/DTOs/Responses/ResponseProductDto.cs
@@ -4,11 +4,11 @@
     {
         public int ProductId { get; set; }
         public required string Name { get; set; }
-        public required double ItemPrice { get; set; }
+        public required decimal ItemPrice { get; set; }
         public required bool IsExempt { get; set; }
         public required bool IsExported { get; set; }
 
-        public ResponseProductDto(int productId, string name, double itemPrice, bool isExempt, bool isExported)
+        public ResponseProductDto(int productId, string name, decimal itemPrice, bool isExempt, bool isExported)
         {
             ProductId = productId;
             Name = name;
