+++ SALESbe/Models/ValueObjects/Product.cs
@@ -5,14 +5,14 @@
         public int ProductId { get; set; }
         public int? Quantity { get; set; }
         public required string Name { get; set; }
-        public required double ItemPrice { get; set; }
-        public double? TotalPrice { get; set; }
+        public required decimal ItemPrice { get; set; }
+        public decimal? TotalPrice { get; set; }
         public required bool IsExempt { get; set; }
         public required bool IsExported { get; set; }
 
         public Product() { }
 
-        public Product(int productId, int quantity, string name, double itemPrice, double totalPrice, bool isExempt, bool isExported)
+        public Product(int productId, int quantity, string name, decimal itemPrice, decimal totalPrice, bool isExempt, bool isExported)
         {
             ProductId = productId;
             Quantity = quantity;
